# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Edna Bar Book Printing API
  description: |
    A comprehensive RESTful API for managing book printing operations.
    
    This API provides complete functionality for:
    - Managing books, authors, and publishers
    - Handling printing companies (private and contracted)
    - Creating and tracking printing contracts
    - Scheduling and monitoring printing jobs
    - Real-time delivery tracking
    
    Built with Go following clean architecture principles.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/edna-bar/api
    email: support@edna-bar.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://edna-bar.com/terms
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.edna-bar.com/api
    description: Production server
...
paths:
...
  /autores:
    get:
      tags:
        - authors
      summary: List all authors
      description: Retrieve a list of all authors in the system
      parameters:
        - name: limit
          in: query
          description: Maximum number of authors to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: name
          in: query
          description: Filter authors by name (partial match)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - authors
      summary: Create a new author
      description: Add a new author to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthorRequest'
      responses:
        '201':
          description: Author created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /autores/{rg}:
    get:
      tags:
        - authors
      summary: Get author by RG
      description: Retrieve a specific author by their RG
      parameters:
        - name: rg
          in: path
          required: true
          description: Author RG
          schema:
            type: string
      responses:
        '200':
          description: Author found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - authors
      summary: Update author
      description: Update an existing author
      parameters:
        - name: rg
          in: path
          required: true
          description: Author RG
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuthorRequest'
      responses:
        '200':
          description: Author updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - authors
      summary: Delete author
      description: Remove an author from the system
      parameters:
        - name: rg
          in: path
          required: true
          description: Author RG
          schema:
            type: string
      responses:
        '204':
          description: Author deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /autores/{rg}/books:
    get:
      tags:
        - authors
      summary: Get author's books
      description: Retrieve all books written by a specific author
      parameters:
        - name: rg
          in: path
          required: true
          description: Author RG
          schema:
            type: string
      responses:
        '200':
          description: Books found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Publishers endpoints
...
  /contratos:
    get:
      tags:
        - contracts
      summary: List all contracts
      description: Retrieve a list of all contracts
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contract'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - contracts
      summary: Create a new contract
      description: Add a new contract to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '201':
          description: Contract created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /contratos/{id}:
    get:
      tags:
        - contracts
      summary: Get contract by ID
      description: Retrieve a specific contract by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Contract ID
          schema:
            type: integer
      responses:
        '200':
          description: Contract found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Printing Jobs endpoints
...
components:
  schemas:
    Book:
      type: object
      properties:
        isbn:
          type: string
          description: Book ISBN
          example: "978-0307474728"
        titulo:
          type: string
          description: Book title
          example: "Cem Anos de Solidão"
        data_de_publicacao:
          type: string
          format: date
          description: Publication date
          example: "1967-06-05"
        editora_id:
          type: integer
          description: Publisher ID
          example: 1
      required:
        - isbn
        - titulo
        - data_de_publicacao
        - editora_id
...
    Author:
      type: object
      properties:
        rg:
          type: string
          description: Author RG
          example: "12345678901"
        nome:
          type: string
          description: Author name
          example: "Gabriel García Márquez"
        endereco:
          type: string
          description: Author address
          example: "América Latina"
      required:
        - rg
        - nome
        - endereco
    CreateAuthorRequest:
      type: object
      properties:
        rg:
          type: string
          description: Author RG
          maxLength: 20
          example: "12345678901"
        nome:
          type: string
          description: Author name
          maxLength: 255
          example: "Gabriel García Márquez"
        endereco:
          type: string
          description: Author address
          maxLength: 255
          example: "América Latina"
      required:
        - rg
        - nome
        - endereco
    UpdateAuthorRequest:
      type: object
      properties:
        nome:
          type: string
          description: Author name
          maxLength: 255
          example: "Gabriel García Márquez"
        endereco:
          type: string
          description: Author address
          maxLength: 255
          example: "América Latina"
...
    Contract:
      type: object
      properties:
        id:
          type: integer
          description: Contract ID
          example: 1
        valor:
          type: number
          format: float
          description: Contract value in currency
          example: 15000.50
        nome_responsavel:
          type: string
          description: Responsible person name
          example: "Maria Silva"
        grafica_cont_id:
          type: integer
          description: Contracted printing company ID
          example: 1
        created_at:
          type: string
          format: date-time
          description: Contract creation date
          example: "2024-01-15T10:00:00Z"
      required:
        - id
        - valor
        - nome_responsavel
        - grafica_cont_id
    CreateContractRequest:
      type: object
      properties:
        valor:
          type: number
          format: float
          description: Contract value in currency
          minimum: 0
          example: 15000.50
        nome_responsavel:
          type: string
          description: Responsible person name
          maxLength: 255
          example: "Maria Silva"
        grafica_cont_id:
          type: integer
          description: Contracted printing company ID
          minimum: 1
          example: 1
      required:
        - valor
        - nome_responsavel
        - grafica_cont_id
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        code:
          type: string
          description: Error code
          example: "NOT_FOUND"
        details:
          type: string
          description: Additional error details
          example: "Book with ISBN 978-0123456789 was not found"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
      required:
        - error
        - code
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid input data"
            code: "BAD_REQUEST"
            details: "ISBN must be a valid format"
            timestamp: "2024-01-15T10:30:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            details: "Book with ISBN 978-0123456789 was not found"
            timestamp: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            details: "An unexpected error occurred"
            timestamp: "2024-01-15T10:30:00Z"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# Global security (can be overridden per endpoint)
# security:
#   - bearerAuth: []